+++ src/index.js
@@ -1,4 +1,4 @@
-import ModalInitiater from "../Scripts/ModalInitiater.js";
+import ModalInitiater from "../areas/modal-initiator/ModalInitiater.js";
 import ButtonSetup from "../Scripts/ButtonSetup.js";
 import NewsPage from "../areas/newspage/index.ts";
